Set up database to connect to DATABASE_URL
We're trying to deploy our app to Heroku.
Heroku provides a postgres database on the address represented by the
ENV variable "DATABASE_URL".

In this commit, if that variable is set, we try to establish a
connection to that url. Otherwise we assume it's a dev environment and
just try to connect to a "felix" database in local.