Unambiguously identify user to be deleted
At /players alongside each player name there is a button for deleting the player.
We found a bug, when there were several players with the same name, when clicking on the button to delete one of them, all users with the same name were deleted.

The problem was the delete implementation, was using the name to identify the players for delete.

This commit makes sure we use a unique id for delete the players.