Add migrations file
When deploying our system to, for instance, Heroku, we need to
make sure the schema of the database contains all the tables
needed by the code.

This PR creates a list of changes/additions (e.g: creating table
users, removing table x, etc) by defining each of those changes in
a .sql file within the migrations folder. It also creates a
migration.rb executable that:

- Finds migrations that have not been applied yet
- Executes them
- Mark them as already applied

Note about the implementation:
---

We're storing the "already-applied-migrations" in a migrations
table in the database itself that contains a file column which
will store the filename of the migration file.